requirements.txt
lightning
pyyaml
timm
torch
torchmetrics
torchvision